static tests: enable type checking by use of mypy
Make the initial set of changes for mypy to pass and run it as
part of ./runtests.sh static.

As part of this, `compat` from the storeapi was removed as it was
dead code had mypy issues on almost every line of its implementation.
