+++ .gitignore
@@ -25,3 +25,4 @@ parts
 stage
 prime
 .vscode
+.mypy_cache
