+++ requirements-devel.txt
@@ -3,6 +3,7 @@ flake8==3.5.0
 pyflakes==1.6.0
 fixtures==3.0.0
 mccabe==0.6.1
+mypy==0.540
 testscenarios==0.5.0
 pexpect==4.2.0
 pyftpdlib==1.4.0
